requirements.txt
langchain-ollama>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
RestrictedPython>=7.0
httpx>=0.27.0
pyyaml>=6.0
